Refactor syntax diagnostic rule modules
