+++ src/application/editor-feedback/syntaxDiagnosticTransferRules.ts
@@ -1,0 +1,36 @@
+import type { UnitParameter } from "../../domain/values/Unit";
+import {
+  isExplicitMacroVariable,
+  parseQuotedStringLiteralContent,
+} from "./syntaxDiagnosticStringValidators";
+
+export const isAbsoluteTransferFilePath = (value: string): boolean =>
+  value.startsWith("/") ||
+  value.startsWith("\\") ||
+  /^[A-Za-z]:[\\/]/.test(value);
+
+export const isValidExplicitTransferFileValue = (
+  parameter: UnitParameter | undefined,
+): boolean => {
+  const rawValue = parameter?.value;
+  if (!rawValue) {
+    return false;
+  }
+
+  return (
+    parseQuotedStringLiteralContent(rawValue) !== undefined ||
+    isExplicitMacroVariable(rawValue)
+  );
+};
+
+export const hasInvalidExplicitTransferSourcePath = (
+  parameter: UnitParameter | undefined,
+): boolean => {
+  const rawValue = parameter?.value;
+  if (!rawValue) {
+    return false;
+  }
+
+  const quotedContent = parseQuotedStringLiteralContent(rawValue);
+  return quotedContent ? !isAbsoluteTransferFilePath(quotedContent) : false;
+};
