+++ src/application/editor-feedback/syntaxDiagnosticTypes.ts
@@ -1,0 +1,11 @@
+export type SyntaxDiagnosticDto = {
+  line: number;
+  column: number;
+  length: number;
+  message: string;
+  severity: "error";
+};
+
+export type BuildSyntaxDiagnosticsOptions = {
+  scheduleLimitYear?: number;
+};
